apps/games: Rename Shift game to Brickmatch
