+++ games/brickmatch/bm_input_joystick.h
@@ -1,5 +1,5 @@
 /****************************************************************************
- * apps/games/shift/shift_input_joystick.h
+ * apps/games/brickmatch/bm_input_joystick.h
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -26,7 +26,7 @@
 
 #include <nuttx/input/djoystick.h>
 
-#include "shift_inputs.h"
+#include "bm_inputs.h"
 
 /****************************************************************************
  * Preprocessor Definitions
